+++ src/app/curves/modes/paint-mode.ts
@@ -6,11 +6,11 @@ export abstract class PaintMode {
   protected readonly predictCanvas: CanvasRenderingContext2D;
   protected readonly mainCanvas: CanvasRenderingContext2D;
   protected settings: Settings;
-  protected lastPointer: Float32Array;
+  protected lastPointer: Uint32Array;
 
-  abstract OnLazyUpdate(lastPointer: Float32Array): void;
-  abstract OnMoveBegin(point: Float32Array, button: number): void;
-  public OnMoveOccur(point: Float32Array, button: number): void {
+  abstract OnLazyUpdate(lastPointer: Uint32Array): void;
+  abstract OnMoveBegin(point: Uint32Array, button: number): void;
+  public OnMoveOccur(point: Uint32Array, button: number): void {
     this.lastPointer = point;
   }
   abstract OnMoveComplete(pointerHasMoved: boolean, button: number): FreeLine | StraightLine;
